requirements.txt
google-generativeai>=0.3.0
markdown>=3.5.0
weasyprint>=60.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytest>=7.4.0
